+++ evm-contracts/src/v0.6/interfaces/AggregatorV3Interface.sol
@@ -3,25 +3,25 @@ pragma solidity >=0.6.0;
 interface AggregatorV3Interface {
   function decimals() external view returns (uint8);
   function description() external view returns (string memory);
-  function getRoundData(uint256 _roundId)
+  function getRoundData(uint80 _roundId)
     external
     view
     returns (
-      uint256 roundId,
+      uint80 roundId,
       int256 answer,
       uint256 startedAt,
       uint256 updatedAt,
-      uint256 answeredInRound
+      uint80 answeredInRound
     );
   function latestRoundData()
     external
     view
     returns (
-      uint256 roundId,
+      uint80 roundId,
       int256 answer,
       uint256 startedAt,
       uint256 updatedAt,
-      uint256 answeredInRound
+      uint80 answeredInRound
     );
   function version() external view returns (uint256);
 }
